App/Creator - + created random treasure generator to generate treasures within Israel. + seperated mock treasure generator button to it's own file
