+++ app/app/creator/index.tsx
@@ -2,8 +2,8 @@ import { Button, StyleSheet, View } from "react-native";
 import TreasuresListView from "../../components/creator/TreasuresListView";
 import TreasuresMapView from "../../components/creator/TreasuresMapView";
 import RoomTitleInput from "../../components/creator/RoomTitleInput";
-import TreasureCreateModalButton from "../../components/creator/TreasureCreateModalButton";
 import { Treasure } from "../../models/Treasure.model";
+import MockTreasureCreateButton from "../../components/creator/MockTreasureCreateButton";
 
 export default function CreatorScreen() {
 	function onNewTreasure(treasure: Treasure): void {
@@ -23,32 +23,6 @@ export default function CreatorScreen() {
 			<MockTreasureCreateButton onNewTreasure={onNewTreasure} />
 		</View>
 	);
-}
-
-type MockTreasureCreateButtonProps = {
-	onNewTreasure: (treasure: Treasure) => void;
-};
-function MockTreasureCreateButton({
-	onNewTreasure,
-}: MockTreasureCreateButtonProps) {
-	function generateRandomTreasure() {
-		return {
-			id: "1a2b3c",
-			name: "Golden Crown",
-			searchRadius: 150,
-			isFound: false,
-			coordinate: {
-				latitude: 34.052235,
-				longitude: -118.243683,
-			},
-		} as Treasure;
-	}
-	function onGenerateTreasure() {
-		const randomTreasure = generateRandomTreasure();
-		onNewTreasure(randomTreasure);
-	}
-
-	return <Button title="Generate Treasure" onPress={onGenerateTreasure} />;
 }
 
 const styles = StyleSheet.create({
